feat(wasm): 6.2 — cross-platform native↔wasm golden (caught 2 real determinism bugs)
Wire the first genuine native↔wasm byte-identity test:
crates/mapgen-wasm/tests/cross_platform.rs runs the full generate pipeline
compiled to wasm32 in Node (`wasm-pack test --node`) and asserts the seed-42
output hashes identical to the native golden (one shared golden file, so a
re-anchor updates both). This guards the multi-scale atlas's load-bearing
promise: a sector regenerated in the browser must match the CLI bit-for-bit.

The contract had never actually been executed cross-platform — and it didn't
hold. The test caught two silent divergences (plus two latent ones):

  1. poisson.rs `gen_range(0..active.len())` sampled over a `usize` range;
     `rand`'s integer sampler draws a different number of RNG bytes for u64
     (64-bit native) vs u32 (wasm32), desyncing the stream from the first
     active pick → a different mesh (3612 vs 3631 points). Fixed by sampling
     an explicit u64 range — a no-op on native (usize *is* u64 there).
  2. climate::band_precip `.exp()` (production path via climate_seasonal):
     std `exp` is the system libm natively but a different libm on wasm; they
     disagree in the last ULP. Routed through fmath::exp — this moved native
     precip onto the shared libm value, re-anchoring both goldens
     (phase2 39afce5e→f2e8f6f9, full 0b29e65c→2c9e034c; aggregate biome
     distribution unchanged, every realism/svg/visual test still green).
  3. Latent (didn't bite seed 42): cultures `.powf`, patch `.hypot` — routed
     through fmath. sqrt left raw: IEEE-754 mandates it correctly-rounded, so
     it's bit-identical across targets (unlike sin/cos/exp/ln/pow).

Preventative: the fmath purity guard (was mapgen-history-only) now scans every
crate's src/, so a future raw transcendental fails `just check` fast. The wasm
runtime golden runs in CI (build-web job) and via the new `just test-wasm`.

Diagnosis method documented in docs/tuning_log.md § Cross-platform determinism;
BACKLOG item marked done.

+++ crates/mapgen-geom/src/poisson.rs
@@ -31,7 +31,15 @@ pub fn poisson_disk_2d<R: RngCore>(
     push(&mut points, &mut active, &mut grid, grid_w, cell_size, seed);
 
     while !active.is_empty() {
-        let idx = rng.gen_range(0..active.len());
+        // Sample over an explicit u64 range, not the `usize` `active.len()`:
+        // `rand`'s integer sampler draws a different number of RNG bytes for
+        // u64 vs u32, so a `usize` range desyncs the stream between 64-bit
+        // native and 32-bit wasm. Widening to u64 makes both targets draw
+        // identically (and is a no-op on native, where usize *is* u64), which
+        // is what keeps the cross-platform golden — and the whole atlas's
+        // "same sector in browser and CLI" promise — byte-identical. See
+        // crates/mapgen-wasm/tests/cross_platform.rs.
+        let idx = rng.gen_range(0..active.len() as u64) as usize;
         let parent = points[active[idx] as usize];
         let mut accepted = false;
